feat: Add conditional rendering for topic deletion and commenting based on user permissions

+++ src/app/homepage/topic/topic.component.ts
@@ -1,10 +1,11 @@
-import { Component, effect, inject, input } from '@angular/core';
+import { Component, computed, effect, inject, input } from '@angular/core';
 import { ITopic } from '../../../interfaces/topic.interface';
 import { CommentComponent } from '../comment/comment.component';
 import { TopicsService } from '../../../services/topics.service';
 import { AddCommentComponent } from '../add-comment/add-comment.component';
 import { TopicProviderService } from './topic-provider.service';
 import { LoggedinUserProviderService } from '../../../services/loggedin-user-provider.service';
+import { RolesService } from '../../../services/roles.service';
 
 @Component({
   selector: 'app-topic',
@@ -19,6 +20,7 @@ export class TopicComponent {
   topicProviderService = inject(TopicProviderService);
   loggedinUserProviderService = inject(LoggedinUserProviderService);
   user = this.loggedinUserProviderService.currentUser;
+  roleService = inject(RolesService);
   showAddComment = false;
 
   constructor() {
@@ -31,10 +33,43 @@ export class TopicComponent {
     this.showAddComment = true;
   }
   onDeleteTopic() {
-    this.topicsService.deleteTopic(this.topic().id).subscribe({
-      error: (error: Error) => {
-        console.error('Error deleting topic:', error);
-      },
-    });
+    if(this.canDeleteTopic()) {
+      this.topicsService.deleteTopic(this.topic().id).subscribe({
+        error: (error: Error) => {
+          console.error('Error deleting topic:', error);
+        },
+      });
+    }
   }
+
+  canDeleteTopic = computed((): boolean => {
+    const userRole = this.roleService.rolesData().find((role) => {
+      return role.id === this.user()?.role;
+    })
+    if(userRole === undefined) return false;
+    const hasRight = !!(userRole?.rights & 4); // Assuming 4 is the bitmask for delete rights
+    const isAdmin = userRole.rights === 15;
+    const isAuthor = this.topic()?.author.id === this.user()?.id;
+    return isAdmin || (hasRight && isAuthor);
+  });
+
+  canComment = computed((): boolean => {
+    const userRole = this.roleService.rolesData().find((role) => {
+      return role.id === this.user()?.role;
+    });
+    if (userRole === undefined) return false;
+    const hasRight = !!(userRole?.rights & 2); // Assuming 2 is the bitmask for comment rights
+    return hasRight;
+  });
+
+  canDeleteComment = computed((): boolean => {
+    const userRole = this.roleService.rolesData().find((role) => {
+      return role.id === this.user()?.role;
+    });
+    if (userRole === undefined) return false;
+    const hasRight = !!(userRole?.rights & 2); // Assuming 4 is the bitmask for delete rights
+    const isAdmin = userRole.rights === 15;
+    const isAuthor = this.topic()?.author.id === this.user()?.id;
+    return isAdmin || (isAuthor && hasRight);
+  });
 }
